[ftl] Reject map pages with all 0xff in the second half of OOB.
This rejects map pages with all 0xff in the second half of the OOB in
order to reject partial page writes which result in wear count listed as
the max value and the validity bits are 0xff, though validity bits are
currently all 0xff.

This also starts writing a form of checksum to the last validity byte to
ensure that the write completed, and to spot any unexpected 0xff
patterns in the oob that may have resulted from an incomplete write.

multiply: corrupt-tests (fuchsia): 20
Bug: b/179400686
Change-Id: I23c56c294ee6eceee28e4cc0a1a26ce0cb7b1880
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/529803
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>