[sdk-tools] Fix handling numeric properties
The properties persisted via `fconfig` are stored via
commandline invocations of `ffx`. The values are then
read by invoking `ffx` and unmarshalling the values via
JSON.

Numeric values are unmarshalled as numbers which was causing
the mapping from the map to data structure to fail, since it
expected string values for all properties.

This fix changes the mapping method to use sprintf to get the
string value for each entry in the map.

Fixed: 76063

Change-Id: I4acb32e68d74b27d59614d5bd17917c579ccfbd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534237
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>