[modular] Don't dispatch TestHarness events on launcher shutdown
TestHarnessFixture tests wait for the TestHarness to shut down
as part of TearDown. This may cause OnNewComponent events to be
dispatched to FakeComponents that were created in the test and
destroyed, or that are using variables in the test scope.

This prevents the FakeComponent OnCreate handler being unexpectedly
invoked after the test is finished. This may happen if the test
restarts Modular, for example.

Also, don't terminate TestHarness when a client disconnects.

Fixed: 77109
Change-Id: I6af0855244d26edea1ddeeb1efa4c2aedfaff557
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533649
Reviewed-by: Gabe Schine <thatguy@google.com>
API-Review: Gary Bressler <geb@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>