[fxfs] Some initial changes to support allocator reservations
We need to be able to reserve space for the journal so that writing to
it should never fail with an out-of-space error condition (for which
we'd never be able to recover from), so we need a way to reserve space
for it.

To support this, I changed associated objects slightly so that they
support both owned an borrowed objects.

This includes changes to the journal such that it will pause new
transactions when waiting for a compaction to finish to relieve space in
the journal.

I've moved the handle out of JournalWriter which means that we no longer
need to hold the writer lock when trimming space at the beginning of the
journal.

Change-Id: I03a1046c969b2e7496a06c4c18c19174a04f4958
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533284
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>