[audio][aml-g12-tdm] Use frame size to adjust mclk
The AMLogic g12 TDM clock configuration needs to change when we use
different frame sizes. This was found to be needed for the PCM output
used for BT HFP support. In BT HFP we use (depending on configuration)
a one channel 16 bits slot PCM configuration with a frame size of 16
bits. This in contrast with I2S that uses 64 bits frames.
In this CL we adjust the mclk configuration based on frame size to get
the correct clocking setting.

Bug: 76759
Test: aml-g12-tdm-test
Test: Logic analyzer check of PCM streams at different frame sizes
(16 and 64 bits) and different frame rates (8, 16, 32, 48 and 96 kHz)
on an Astro "Oportus Edition" with PCM pins exposed.
Test: Manual audio check on Astro and Sherlock.

Change-Id: I90038733befe9dd9fccb770465808ca34cf5f2d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532061
Reviewed-by: Eric Holland <hollande@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>