[time] Update frequency documentation.
The original algorithm was a bit inconsistent with the handling of
frequency, occasionally working with utc per monotonic and occasionally
working with monotonic per realtime. The implementationation is
consistent in only using utc per monotonic and this CL cleans up the
documentation to match and be more explicit about its use of frequency.

I'm also removing the note that frequency isn't implemented yet to
celebrate the fact that we finished it!

Bug: 56868
Change-Id: I093ce8f5dc9c5b83f380ebbbf933c578b440bdb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534981
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>