[sdk-tools] fssh, fserve respect configured sshport
Updates fssh and fserve to check for a custom ssh port being
configured via `fconfig` if the --device-ip flag is not used.

The device-name is used as the key to read fconfig values, so
if --device-ip is given, there is no way to look it up.

The method RunSSHCommandWithPort is a transitional method to allow
migration of other tools that use this library in other repos. This
method will be renamed RunSSHCommand when the migration is completed.

Fixed: 77410
Change-Id: Ibb51a0f7a1cd8157fcbb54f485028c3ed7a19842
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534885
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>