[kernel][x86] Remove `amd_retpoline` global
This value gave whether we were using the AMD-appropriate retpoline
patch alternative (printed in `x86_feature_debug()`). The code-patching
infrastructure now prints this information out on boot instead and
nothing sets this anymore. Given that we do log the relevant information
(and that it would be a bit awkward to go back to setting this variable
in feature.cc), we remove it in this change (as was originally meant to
be done).

Bug: 77306
Change-Id: I75459d777119a79e86a0029d8fcfe15739152169
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535381
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>