[kernel_cmdline] Disable virtio_ethernet in GCE
GCE VMs are experiencing kernel lockups due to a bad interaction between
the virtio_ethernet driver and Vanadium, see the attached bug for more
context. Note that this will break the core.x64-gce builder, but we
cannot pass this in from the builder definition because we need the
kernel arg in the image.

Bug: b/188675754
Change-Id: I47dbafeeede9c72ab118be499d8bb85a94df60ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534682
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>