[display] Add flag to disable kernel framebuffer.
Forcing all virtcon images to be set as kernel framebuffer
limits that formats that can be used and significantly
impacts performance when virtcon is not in single framebuffer
mode.

This adds a flag that disable kernel framebuffer support
for configurations where that is desirable.

Bug: 76124
Test: display-test (with driver.display.disable-kernel-framebuffer=1)
Change-Id: I4df0819cc3d460ae3091e99d9c123a8f9d25cdb0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/527241
Reviewed-by: Arman Uguray <armansito@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: David Reveman <reveman@google.com>