[ffx] manual targets never enter disconnected
A manual target is non-discoverable. The disconnected state is
essentially a mis-nomer that actually means "needs to be rediscovered",
as actions such as connecting to the target, or listing in target
listings are dependent on the state being not-disconnected.

update_connection_state enforces this transition.

Bug: 76632
Fixed: 76426
Change-Id: I32463c6bf8f1ab53a5f0cf36fe3aaf5016b0920c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534363
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>