[stress_tests] Make file names unique
When randomly generating file names there's a chance of generating
duplicate file names close together which is suspected to be the cause
of fxbug.dev/77326. This cl changes the file name generation to be
based on a u64 counter which guarantees file name uniqueness.

Multiply: minfs-short: 100
Fixed: 77326
Change-Id: I3a077ac4af2c37ed8f5858b41f669bd6eb14555d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535825
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>