[fx] Add fvdl remote --streamed=false
fxr/527780 added the `fvdl remote` subcommand with no functionality.

This CL adds the first functionality for that subcommand by enabling
the unstreamed variant. This variant calls into a remote workstation
to perform fuchsia compilation, then copies the binaries to the local
machine for execution and launches a local instance of the emulator.

Change-Id: I3db4e213a7edb27553604193a32c1c75b4d14db3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533342
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>