[cmc] Support offering/exposing multiple capabilities together
Add array syntax for exposing or offering multiple capabilities in a
bundle.

This was already supported for services, protocols, events. The change
extends this to other capability types.

Note that we aren't introducing this syntax for use/capability. That's
because those declarations require unique options so there isn't as much
value in bundling them together.

Change-Id: Ie9daf1092bbcd7c4518e7cd34554a057f15eaee8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531045
Reviewed-by: Justin Mattson <jmatt@google.com>
Commit-Queue: Gary Bressler <geb@google.com>