[fidl][summary] Change API signatures
Reviewers note: the only real change is in `fidl_library.gni`.  The
`.api` file changes are all changes of the library hash value.

The API signatures are now computed based on the API summary files, not
the normalized form of the library.  This should be a more stable format
for API summaries, changing the API signature only when real API
impacting changes happen.

As result of the change of the signature method, all API signatures are
changing.  This is inevitable, and may need to be undone in case this
change backfires somehow.

We'll start checking in the API summary files alongside the respective
APIs, but for API summary checks this is not yet needed.  No parts of
prior fixtures are removed yet, everything is flag-guarded.

Bug: 68952
Change-Id: If026d7a427ed7fe577e98be2426267aafb0622db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533141
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
API-Review: Pascal-Louis Perez <pascallouis@google.com>