[fxfs] Change compaction strategy
Prior to this change, we would always compact all layers but that had
the potential to exhaust the journal, so this change introduces a limit
for object store compactions so that we will only compact some of the
layers to keep below a given limit.

This fixes the DirectoryTest.Large/Fxfs test which got broken with a
previous change.

Change-Id: I1184c755c71361cee4dce292723f123e84a71dd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534021
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>