[fvm] Remove snapshot metadata format
After this change, an FVM partition added with --reserve-slices will be
zeroed, rather than being populated with the snapshot metadata which was
going to be used for the A/B snapshots feature.

This metadata is never loaded by the FVM driver, nor is it read by the
FVM host tooling, so this change has no functional effect.

Tested: Built + Ran.
Bug: 77596
Change-Id: Ie6addf86cf472b8a64afd88a9589344a81bc8772
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535826
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>