[fxfs] Sequence Items in the tree
Adds a monotonically increasing sequence number to Item which can be
used for a number of purposes:

- Filtering out items which were inserted while a graveyard purge was
  occurring, so that we only purge items that were already in the
  graveyard at the time of starting a purge.
- Filtering out directory entries which are added while we're sending
  the initial state of the directory in the Watcher API.
- Debugging.

The sequence value is initialized to 0, and set to the current journal
offset at the time of committing a transaction (in
Mutations::apply_mutations). During merging, the common case of EmitLeft
or Keep preserves sequences; for the more interesting cases where a new
item is emitted, we use the minimum of the two items' sequences, since
the sequence number is most useful when it represents the "birth" of
some object. This only comes into play for extents and allocations.

Tested: Unit.
Change-Id: I6eea7cfc1d7f7171f2eda0d254ca17518bfc2696
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532301
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>