[virtualization][x86] Add basic tests to the I/O APIC.
Add some very basic tests exercising the I/O APIC. The tests exercise
register reading/writing, but don't attempt to exercise the actual
mechanism for sending interrupts to the guest.

In the process, fix a minor bug where the I/O APIC was expecting 32-bit
writes to its IND register, instead of 8-bit writes as specified in the
spec. This bug didn't affect anything in practice, because calling code
always initialised the upper bits to 0, and the little-endian layout
meant that 8-bit and 32-bit interpretations of the IoValue type produced
the same result.

Change-Id: I8837cf4f410c4f4df2248bd80b5f52124229f2f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536628
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>