[cm][routing] Make directory rights tests generic
Defines a CommonRightsTest providing tests of routing
of directory capabilities with restricted rights and
moves a subset of the existing routing-with-rights
unit tests into it. Runs these tests for both
ComponentManager and for the static routing analyzer.

The remaining tests will be run for the static analyzer
once it has support for framework capabilities.

Bug: 61861
Change-Id: I4fbd3d72f9e69b02dbd51ef680e9af4f2e276031
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523861
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Laura Peskin <pesk@google.com>