[mold] Fixed rare partial updates bug.
The recursive unique element counting alrogithm implemented to take
advantage of the sorted array was incorrect, but this was hard to
notice since it worked in most cases and tended to err on the
conservative side.

This CL introduces a fix for it and, as a bonus, replaces a last
cover emptiness check.

Testing: fx test mold-tests && fx test surpass-tests
Change-Id: Ia2d4a68cff9018f415f4afbff0c770cc39000c4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537061
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: David Reveman <reveman@google.com>