[input][shortcut] Support press-release shortcuts
It turned out that shortcuts that involve a press and a release of
a single (modifier) key didn't work because of a bug.

This change makes it possible to use such shortcuts, and adds
new tests to guard the functionality.

While I was there, reorganized a few things:
- changed BUILD.gn related rules to be next to each other,
  and test fixture names to be a little less similar.
- added `Latch` and pulled handled assertions outside of assert hooks,
  thus allowing assertions to fire multiple times with different
  "handled" states.  This allows us to observe the end state
  of the test instead each individual state.
- Renamed the shortcut manager package used in the hermetic tests.

Fixed: 76252
Change-Id: I42bcd9c5d7d1c1e6db3ebfc5e220c43285e245d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526517
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>