[packet][records] Overhaul options LengthEncoding
Previously, `OptionsImplLayout` had two const parameters,
`LENGTH_ENCODING` and `OPTION_LEN_MULTIPLIER`, which had interlocking
correctness properties - certain pairs of values for these parameters
were invalid, and validity was enforced via a hack which would fail at
link time if illegal pairs of values were used.

The `LENGTH_ENCODING` parameter had the following type:

    pub enum LengthEncoding {
        TypeLengthValue,
        ValueOnly,
    }

This commit removes the `OPTION_LEN_MULTIPLIER` parameter, and redefines
`LengthEncoding` like so:

    pub enum LengthEncoding {
        TypeLengthValue { option_len_multiplier: NonZeroUsize },
        ValueOnly,
    }

This new definition makes it impossible to represent the illegal pair of
values that was a possibility before.

This commit also fixes a bug in options parsing in which, if
`LENGTH_ENCODING` were set to `ValueOnly`, the parser would still treat
the length field as encoding the length of the kind and length bytes
along with the body bytes.

Fixed: 73253

Test: fx test packet-test test_length_encoding

Change-Id: I92048be39df7bbd74307b15f1e96b1f375d4bbfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507000
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Marina Ciocea <marinaciocea@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>