[pkg-cache] Model the retained package index
This change adds a new index::retained::RetainedIndex type that the
pkg-cache will use to track the state of the retained package index. A
future change will define a new index type that composes the dynamic and
retained indices into a single index for pkg-cache to interact with.

Fixed: 75924
Change-Id: Id227660cedaa58c6cbef58a53ab69d5b95ed4c9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537005
Commit-Queue: Kevin Wells <kevinwells@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>