[starnix] Pass PipeTest.Seek
This test requires lseek to return ESPIPE for pipes. We still do not
implement the actual seek logic for regular files, but this CL prepares
us for doing that once we have a test that requires it.

Also, update the test filter to run more tests that we already pass.

Bug: 76861
Bug: 77376
Change-Id: I6e98ac9e44b42e5944bc621b75470e155d7900ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537761
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>