[fidlconv] Cache derived resourceness during compilation
When converting method request/response types, fidlconv needs to
know whether the top level type is a resource in order to annotate
it correctly in the new method syntax.

Previously this was done in a fragile way by attempting to reconstruct
fidlc's resourceness logic.
This CL updates fidlc to store derived resourceness for request/response
structs, which fidlconv can then use to perform lookup on during
conversion instead of trying to recompute the resourceness.

This approach breaks abstraction somewhat by exposing Library internals
to fidlconv, but this is considered OK given it is only temporary.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler
--gtest_filter='ConverterTests.AliasOfResource'

Change-Id: I8716225b5ee4d062a5aec44c86e2450b4c0387d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536938
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Felix Zhu <fcz@google.com>