[system-updater] Reland Migrate system-updater to a v2 component.
This change migrates the system-updater component from the software
delivery stack into the core realm as part of the v1-v2 migration
project. System-updater is an eagerly started component that provides
access to the fuchsia.update.installer.Installer service.

As part of this migration the component depends on several services
provided by other components in the SWD stack. These are routed through
core.cml to the component. In addition to this this component relies on
a set of features & directories:
- build-info
- deprecated-misc-storage
- isolated-persistence-storage
- pkgfs/system

Relanding this change moves the system-updater routing in core into a
shard so that it does not have to be included in recovery images. These
don't require the updater and are space limited and cannot support the
eager requirement of the component.

Bug: 73992
Change-Id: I1893e4b8cb06470d5ab7d591ee74590000d1f479
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534309
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>