[fxfs] Verify checksums when replaying the journal
When replaying the journal, we can't be sure that data reference by
transactions has been written, so use checksums to validate mutations.

This also changes writes so that they are written more concurrently
but there is still more that could be done.

This change also required me to fix tombstone so that it breaks up
transactions to prevent them from getting too big.

Change-Id: Ia60613d9cf2bf6ade1695f720c3eeedc8ab2130e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537401
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Suter <csuter@google.com>