[pkg-resolver] Cache TUF metadata with non-auto repos
Previously, we would try to update TUF every time before we try to
resolve a package. This can generate a significant amount of load during
an OTA, since we would update for each package. This patch fixes this by
allowing the TUF metadata to be cached for 5 minutes, which should
significantly reduce the load on the server.

Fixed: 77442
Change-Id: I70b5a5682a56dd706f26b6536bed13565f6418bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538013
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>