[scenic][ui-input] f.u.pointer: Combine win with first message in exclusive mode
Makes it so that when injecting with exclusive mode the initial message
delivered to the fup client includes the "win" message for the
interaction. Before this the win would be sent separately, usually after
another hanging-get interaction.

The corresponding fix for the hit tested mode will come in a follow-up
CL.

Bug: 76233
Test: fx test input_unittests
Change-Id: If88ce4359028cd59c27e6e9a933239fc73be4df6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536911
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>