[lockdep][kasan] De-dupe CUSTOM_DEFAULT_STACK_SIZE defines
When using both `--variant kasan` and `--args "enable_lock_dep=true"`
CUSTOM_DEFAULT_STACK_SIZE is defined twice via separate GN config()
blocks.

There did not appear to be a precise calculation of these values, and
each was 2x and 1.5x the default respectively, so taking the max of the
two and confirming experimentally that this is a sufficient value when
both options are used.

Tests: fxrev.dev/537835 based on top of this CL, enables lockdep by
default and we successfully build && boot, however kasan finds an issue
in lockdep code.

Bug: 30033, 72799

Change-Id: I622491d74cb9450a4455a9a6af38dae7d89dc8e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537834
Commit-Queue: Nathan Mulcahey <nmulcahey@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>