[forensics] Add internal FIDL protocol for migration data
Feedback's components are being combined into a single monolith that
will serve all the Feedback protocols. The data in namespaces of the
original components will need to be migrated into the namespace of the
new component in order to preserve persistent state.

This protocol will be internal to Feedback (and not in the SDK) and
short lived (it will be deleted once the merge is complete).

Bug: 71830

Change-Id: I7eee5bfba77bcd66cdf3c2a121ce4a70aed2c893
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536908
API-Review: Gary Bressler <geb@google.com>
Commit-Queue: Alex Pankhurst <pankhurst@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>