[zbitl] ignore/take View error before exiting scope
This change updates several uses of zbitl::View in the tree in which
we forget to `ignore_error()` before exiting for a non-iteration
failure reason. In not doing this, we risk crashing a program in which
we could have otherwise more gracefully reported an error.

Change-Id: Ic34bd61500d35ad0055221c92885101267acd9f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537305
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>