[netstack/tests] Migrate inspect integration test to v2
And make the necessary corresponding changes to bits of the common
integration test library that are exercised by this test. Also, add a
client CML shard for the netemul-v2 sandbox to D.R.Y. the creation of v2
manifests for integration tests that use the sandbox.

Bug: 77202
Fixed: 77458
Test: fx test -vdo netstack-integration-tests-v2
Multiply: netstack-integration-tests-v2
Change-Id: I2ab85ad70360a322662a32f74333d19e6c1ba32e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/531986
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>