[brcmfmac] Fix dropped scan results race condition
When starting a scan we associate a sync ID with each scan, this is sent
to firmware and each scan result contains this sync ID to ensure that
the scan result belongs to the currently ongoing scan. The sync ID is
not stored until the scan has started, this means that it's vulnerable
to a race condition where if the thread that starts the scanning gets
suspended before it stores the sync ID results could arrive and they
would be matched against an incorrect sync ID. This would cause the
results to be dropped.

This fixes the issue by introducing a lock such that the whole process
of starting a scan completes before we start processing results.

Test: Manual testing able to reproduce without fix, not with fix
Bug: fxbug.dev/77630
Change-Id: I6dbe8d772214fe8ed58f785ba899dfa24ab2b0ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536032
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>