[botanist] Run subprocess without piping stdin
Botanist uses the subprocess golib in tools/lib/subprocess to run
subprocesses. This library installs a handler that SIGKILLs all child
processes when a SIGTERM is received; it accomplishes this by sending a
SIGKILL to a PGID. Unfortunately, https://fxrev.dev/498141 breaks this
behavior because it does not set a PGID when stdin is being rerouted
from os.Stdin. This led to cases where botanist would get a SIGTERM and
stop the serial server, but testrunner would continue to run. This led
to the attached flake. This CL gets around this issue by removing
os.Stdin piping to testrunner, which isn't actually necessary.

Bug: 72696

Change-Id: I9edfe7269390f4b3bda832e5f031de55aba6ce3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538989
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>