[fidlfmt] Create new fidl formatter
This is the first commit of the fidlfmt rewrite for the new FIDL syntax.
It only supports formatting library and using declarations at the
moment, though the basic scaffolding for supporting pretty printing of
other raw AST nodes in a "visitation" manner similar to that used by
fidlconv is present.

Test: fidldev test fidlc
Bug: 73507
Change-Id: Id54e6e0ea5209bc0c0f82f49c8f76886db001a9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534103
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>