[fidlc][lsc] Make handle rights flag a no-op
This CL enables the "enable_handle_rights" flag always, by making the
actual flag a no-op. This means the remaining usages of the flag can
be deleted.

Handle rights works in all bindings now and with the exception of Dart
GIDL tests is functionally complete. Handle rights is no longer an
experimental feature and it seems like the appropriate time to quietly
enable the functionality in anticipation of a more visible launch.

fx lsc presubmit was run on this CL

Bug: 77846
Bug: 41920
Change-Id: Idff49e42a07fe6c1e61b0d543313220a32f1f6ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537461
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>