[debugger] Implicitly find variables on const this.
Previously the code did not strip C-V qualifiers when computing the
collection type for the "this" pointer. This meant that attempting to
print variables on "this" without qualification for const functions
failed. The situation is particularly noticeable for printing captured
variables in lambdas, which are const by default.

Bug: 77910
Change-Id: If36fa3ef764194748add0e0b11f8bcdd9c3bee5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539021
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>