[blobfs][test] Deflake BlobfsFragmentationTest
The test was checking for specific metric values from Cobalt,
but blobfs does the metrics flushing to Cobalt in a background thread,
and it is not guaranteed that all of the metrics are updated atomically,
so partial results were possible.

Changed the test to wait until the metrics were the expected value, on a
deadline, which should help reduce flake.

Bug: 75799
Change-Id: I326a2987ceab8a917ee1ee1f19c2a02fbff42f20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522521
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>