[scenic] Fix flatland-display example for femu.
The example app was failing on femu due to using `vmo.write()`, which
is not supported for physical VMOs.  Instead, this CL first maps the
VMO, then writes via the mapped ptr.

Note: not all platforms are guaranteed to have a sysmem format which:
- is CPU accessible
- supports Vulkan sampling
- supports direct display scan-out

... the proper solution is probably to use Vulkan to generate the
client image.  This would avoid needing to allocate a CPU-accessible
buffer.

Change-Id: Ia07d01564ffe76e8f8e8d5c1cbc2351d349308a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536265
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Dragos Tiselice <dtiselice@google.com>