[build][go] Add sdk_name option.
The go_binary template forwards variables to go_build,
which forwards it to sdk_host_tool. go_binary supports
sdk_category but not sdk_name.

We would like to export the fidlgen_llcpp (go_binary) tool to
the SDK under a different name, but keep the original name
in-tree. The existing templates support an output_name field,
but that would also change the in-tree name of the tool.

Bug: 74665
Change-Id: I105cef861026f81100867a4fdb3c56e8ae5546d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537112
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>