[build] Remove non-chromebook-x64 vboot image
This seems to have been introduced at a time in which we had
"chromebook-x64" developers working off the SDK, which was only based
off of a core.x64 build: hence the need to smuggle in a chromebook-x64
image through the latter. That is no longer the case today, as the SDK
includes chromebook-x64 images directly.

Change-Id: I69062ed439ebda07aa84137a96a90c1018b76663
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539063
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>