[phys][page-table] Add support for specifying cache attrs on ARM64.
In particular, we allow regions to be mapped in as either cached or as
device memory. We additionally add support for setting up the ARM64 MAIR
register.

The infrastructure supports more types of cache attributes, but we don't
have a need for these yet.

The implementation requires that users of the library set up the MAIR
register to a library-defined value. This constraint will be lifted in
future if necessary.

Bug: 67632
Change-Id: Iac267c5c7039566d19ef437d3fa31368efe7fd3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533403
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>