[lowpan-spinel-driver] Added cleanup mechanism to API lock
Previously, if an operation like a join or form were interrupted,
then there was no way to put the NCP back in a known state (because
the future controlling that process had been dropped). This change
creates a new type for the API lock that allows us to register
NCP requests that are to be sent when the future is dropped, allowing
us to return the NCP to a known state.

Test: fx test lowpan-spinel-driver-test
Change-Id: Ia0663339acb4ead79f8f18d48d14b30e776d3a69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536161
Fuchsia-Auto-Submit: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>